Upd: DB & Entities refactoring

+++ src/main/java/ru/pavlinina/ecommerce/controllers/AdminController.java
@@ -16,32 +16,32 @@ import ru.pavlinina.ecommerce.services.UserService;
 @RequestMapping("admin")
 public class AdminController {
 
-    @Autowired
-    private UserService userService;
+  @Autowired
+  private UserService userService;
 
 
-    @GetMapping("index")
-    public String index() {
-        return "admin/index";
-    }
+  @GetMapping("index")
+  public String index() {
+    return "admin/index";
+  }
 
-    @GetMapping("user-list")
-    public String userList(Model model) {
-        model.addAttribute("userList", userService.findAllUser());
-        return "admin/user-list";
-    }
+  @GetMapping("user-list")
+  public String userList(Model model) {
+    model.addAttribute("userList", userService.findAllUser());
+    return "admin/user-list";
+  }
 
-    @GetMapping("delete-user/{userId}")
-    public ModelAndView deleteUser(@PathVariable("userId")String userId) {
-        ModelAndView mv = new ModelAndView("admin/user-list");
-        userService.deleteUser(Long.parseLong(userId));
-        mv.addObject("userList", userService.findAllUser());
-        return mv;
-    }
+  @GetMapping("delete-user/{userId}")
+  public ModelAndView deleteUser(@PathVariable("userId") String userId) {
+    ModelAndView mv = new ModelAndView("admin/user-list");
+    userService.deleteUser(Integer.parseInt(userId));
+    mv.addObject("userList", userService.findAllUser());
+    return mv;
+  }
 
-    @GetMapping("add-user")
-    public String addUser() {
-        return "admin/add-user";
-    }
+  @GetMapping("add-user")
+  public String addUser() {
+    return "admin/add-user";
+  }
 
 }
